+++ pkg/vvm/impl_cfg.go
@@ -5,6 +5,7 @@
 package vvm
 
 import (
+	"net"
 	"os"
 
 	"github.com/voedger/voedger/pkg/appdef"
@@ -52,6 +53,8 @@ func NewVVMDefaultConfig() VVMConfig {
 			return mem.Provide(coreutils.MockTime), nil
 		},
 		SecretsReader: isecretsimpl.ProvideSecretReader(),
+		IP:            net.IPv4(127, 0, 0, 1),
+		NumVVM:        1,
 	}
 	if coreutils.IsTest() {
 		res.SecretsReader = itokensjwt.ProvideTestSecretsReader(res.SecretsReader)
