+++ pkg/vvm/wire_gen.go
@@ -19,7 +19,6 @@ import (
 	"github.com/voedger/voedger/pkg/bus"
 	"github.com/voedger/voedger/pkg/coreutils"
 	"github.com/voedger/voedger/pkg/coreutils/federation"
-	"github.com/voedger/voedger/pkg/elections"
 	"github.com/voedger/voedger/pkg/extensionpoints"
 	"github.com/voedger/voedger/pkg/goutils/logger"
 	"github.com/voedger/voedger/pkg/iauthnz"
@@ -68,7 +67,6 @@ import (
 	"slices"
 	"strconv"
 	"strings"
-	"time"
 )
 
 // Injectors from provide.go:
@@ -245,14 +243,14 @@ func ProvideCluster(vvmCtx context.Context, vvmConfig *VVMConfig) (*VVM, func(),
 
 func ProvideVVM(vvmCfg *VVMConfig) (voedgerVM *VoedgerVM, err error) {
 	vvmCtx, vvmCtxCancel := context.WithCancel(context.Background())
-	problemCtx, problemCtxCancel := context.WithCancelCause(context.Background())
+	problemCtx, problemCtxCancel := context.WithCancel(context.Background())
 	vvmShutCtx, vvmShutCtxCancel := context.WithCancel(context.Background())
 	servicesShutCtx, servicesShutCtxCancel := context.WithCancel(context.Background())
 	monitorShutCtx, monitorShutCtxCancel := context.WithCancel(context.Background())
 	shutdownedCtx, shutdownedCtxCancel := context.WithCancel(context.Background())
 	voedgerVM = &VoedgerVM{
 		vvmCtxCancel:                 vvmCtxCancel,
-		clusterSize:                  vvmCfg.ClusterSize,
+		numVVM:                       vvmCfg.NumVVM,
 		ip:                           vvmCfg.IP,
 		problemCtx:                   problemCtx,
 		problemCtxCancel:             problemCtxCancel,
@@ -292,142 +290,6 @@ func ProvideVVM(vvmCfg *VVMConfig) (voedgerVM *VoedgerVM, err error) {
 	return voedgerVM, nil
 }
 
-// deprecated
-func (vvm *VoedgerVM) Shutdown() {
-	vvm.vvmCtxCancel()
-	vvm.ServicePipeline.Close()
-	vvm.vvmCleanup()
-}
-
-// ShutdownNew shuts down the VVM process.
-func (vvm *VoedgerVM) ShutdownNew() error {
-
-	vvm.vvmShutCtxCancel()
-
-	<-vvm.shutdownedCtx.Done()
-
-	select {
-	case err := <-vvm.problemErrCh:
-		return err
-	default:
-		return nil
-	}
-}
-
-// LaunchNew launches the VVM process.
-// leadershipAcquisitionDuration - default 120 sec
-func (vvm *VoedgerVM) LaunchNew(leadershipAcquisitionDuration time.Duration) context.Context {
-	go vvm.shutdowner()
-
-	if err := vvm.tryToAcquireLeadershipInLoop(leadershipAcquisitionDuration); err != nil {
-		vvm.updateProblem(err)
-		return vvm.problemCtx
-	}
-
-	if err := vvm.ServicePipeline.SendSync(ignition{}); err != nil {
-		logger.Error(err)
-		vvm.updateProblem(err)
-		return vvm.problemCtx
-	}
-
-	return vvm.problemCtx
-}
-
-// shutdowner is a routine that waits for the VVM to be shut down.
-func (vvm *VoedgerVM) shutdowner() {
-
-	<-vvm.vvmShutCtx.Done()
-
-	vvm.servicesShutCtxCancel()
-
-	vvm.vvmCtxCancel()
-	vvm.ServicePipeline.Close()
-
-	vvm.vvmCleanup()
-
-	vvm.monitorShutCtxCancel()
-	vvm.monitorShutWg.Wait()
-
-	vvm.electionsCleanup()
-
-	vvm.shutdownedCtxCancel()
-}
-
-// leadershipMonitor is a routine that monitors the leadership context.
-func (vvm *VoedgerVM) leadershipMonitor(leadershipAcquisitionDuration time.Duration) {
-	defer vvm.monitorShutWg.Done()
-
-	select {
-	case <-vvm.leadershipCtx.Done():
-
-		go vvm.killerRoutine(leadershipAcquisitionDuration)
-		vvm.updateProblem(ErrLeadershipLost)
-	case <-vvm.monitorShutCtx.Done():
-		return
-	}
-}
-
-// killerRoutine is a routine that kills the VVM process after a quarter of the leadership duration
-func (vvm *VoedgerVM) killerRoutine(leadershipAcquisitionDuration time.Duration) {
-	time.Sleep(leadershipAcquisitionDuration / 4)
-	os.Exit(1)
-}
-
-// tryToAcquireLeadershipInLoop tries to acquire leadership in loop
-func (vvm *VoedgerVM) tryToAcquireLeadershipInLoop(leadershipAcquisitionDuration time.Duration) error {
-	ttlStorage := storage.NewElectionsTTLStorage(vvm.VVMAppTTLStorage)
-	elections2, electionsCleanup := elections.Provide(ttlStorage, vvm.ITime)
-	vvm.electionsCleanup = electionsCleanup
-
-	timerCh := vvm.ITime.NewTimerChan(leadershipAcquisitionDuration)
-
-	vvm.startedLeadershipAcquisition <- struct{}{}
-
-	vvmIdx := 1
-	for {
-		select {
-		case <-timerCh:
-			return ErrVVMLeadershipAcquisition
-		default:
-
-			leadershipCtx := elections2.AcquireLeadership(TTLStorageImplKey(vvmIdx), vvm.ip.String(), leadershipAcquisitionDuration)
-			if leadershipCtx != nil {
-				vvm.leadershipCtx = leadershipCtx
-				vvm.monitorShutWg.Add(1)
-				go vvm.leadershipMonitor(leadershipAcquisitionDuration)
-				return nil
-			}
-
-			vvmIdx++
-			if vvmIdx > int(vvm.clusterSize) {
-				vvmIdx = 1
-			}
-		}
-	}
-}
-
-// updateProblem writes a critical error into problemErrCh exactly once
-// and sets the cause on problemCtx. This ensures no double-writes of errors.
-func (vvm *VoedgerVM) updateProblem(err error) {
-
-	vvm.problemCtxCancel(err)
-
-	vvm.problemCtxErrOnce.Do(func() {
-		vvm.problemErrCh <- err
-	})
-}
-
-// deprecated
-func (vvm *VoedgerVM) Launch() error {
-	ignition2 := ignition{}
-	err := vvm.ServicePipeline.SendSync(ignition2)
-	if err != nil {
-		err = errors.Join(err, ErrVVMLaunchFailure)
-		logger.Error(err)
-	}
-	return err
-}
-
 func provideIVVMAppTTLStorage(prov istorage.IAppStorageProvider) (storage.IVVMAppTTLStorage, error) {
 	return prov.AppStorage(appdef.NewAppQName(istructs.SysOwner, "vvm"))
 }
@@ -451,7 +313,7 @@ func provideSchedulerRunner(cfg schedulers.BasicSchedulerConfig) appparts.ISched
 	return schedulers.ProvideSchedulers(cfg)
 }
 
-func provideBootstrapOperator(federation2 federation.IFederation, asp istructs.IAppStructsProvider, time2 coreutils.ITime, apppar appparts.IAppPartitions,
+func provideBootstrapOperator(federation2 federation.IFederation, asp istructs.IAppStructsProvider, time coreutils.ITime, apppar appparts.IAppPartitions,
 	builtinApps []appparts.BuiltInApp, sidecarApps []appparts.SidecarApp, itokens2 itokens.ITokens, storageProvider istorage.IAppStorageProvider, blobberAppStoragePtr iblobstoragestg.BlobAppStoragePtr,
 	routerAppStoragePtr dbcertcache.RouterAppStoragePtr) (BootstrapOperator, error) {
 	var clusterBuiltinApp btstrp.ClusterBuiltInApp
@@ -472,7 +334,7 @@ func provideBootstrapOperator(federation2 federation.IFederation, asp istructs.I
 		return nil, fmt.Errorf("%s app should be added to VVM builtin apps", istructs.AppQName_sys_cluster)
 	}
 	return pipeline.NewSyncOp(func(ctx context.Context, work pipeline.IWorkpiece) (err error) {
-		return btstrp.Bootstrap(federation2, asp, time2, apppar, clusterBuiltinApp, otherApps, sidecarApps, itokens2, storageProvider, blobberAppStoragePtr, routerAppStoragePtr)
+		return btstrp.Bootstrap(federation2, asp, time, apppar, clusterBuiltinApp, otherApps, sidecarApps, itokens2, storageProvider, blobberAppStoragePtr, routerAppStoragePtr)
 	}), nil
 }
 
@@ -647,9 +509,9 @@ func provideSubjectGetterFunc() iauthnzimpl.SubjectGetterFunc {
 	}
 }
 
-func provideBucketsFactory(time2 coreutils.ITime) irates.BucketsFactoryType {
+func provideBucketsFactory(time coreutils.ITime) irates.BucketsFactoryType {
 	return func() irates.IBuckets {
-		return iratesce.Provide(time2)
+		return iratesce.Provide(time)
 	}
 }
 
@@ -904,8 +766,8 @@ func provideBlobAppStoragePtr(astp istorage.IAppStorageProvider) iblobstoragestg
 	return new(istorage.IAppStorage)
 }
 
-func provideBlobStorage(bas iblobstoragestg.BlobAppStoragePtr, time2 coreutils.ITime) BlobStorage {
-	return iblobstoragestg.Provide(bas, time2)
+func provideBlobStorage(bas iblobstoragestg.BlobAppStoragePtr, time coreutils.ITime) BlobStorage {
+	return iblobstoragestg.Provide(bas, time)
 }
 
 func provideRouterAppStoragePtr(astp istorage.IAppStorageProvider) dbcertcache.RouterAppStoragePtr {
@@ -996,5 +858,6 @@ func provideServicePipeline(
 	publicEndpoint PublicEndpointServiceOperator,
 	appStorageProvider istorage.IAppStorageProvider,
 ) ServicePipeline {
-	return pipeline.NewSyncPipeline(vvmCtx, "ServicePipeline", pipeline.WireSyncOperator("internal services", pipeline.ForkOperator(pipeline.ForkSame, pipeline.ForkBranch(opQueryProcessors), pipeline.ForkBranch(opCommandProcessors), pipeline.ForkBranch(opBLOBProcessors), pipeline.ForkBranch(pipeline.ServiceOperator(opAsyncActualizers)), pipeline.ForkBranch(pipeline.ServiceOperator(appPartsCtl)), pipeline.ForkBranch(pipeline.ServiceOperator(appStorageProvider)))), pipeline.WireSyncOperator("admin endpoint", adminEndpoint), pipeline.WireSyncOperator("bootstrap", bootstrapSyncOp), pipeline.WireSyncOperator("public endpoint", publicEndpoint))
+	return pipeline.NewSyncPipeline(vvmCtx, "ServicePipeline", pipeline.WireSyncOperator("internal services", pipeline.ForkOperator(pipeline.ForkSame, pipeline.ForkBranch(opQueryProcessors), pipeline.ForkBranch(opCommandProcessors), pipeline.ForkBranch(opBLOBProcessors), pipeline.ForkBranch(pipeline.ServiceOperator(opAsyncActualizers)), pipeline.ForkBranch(pipeline.ServiceOperator(appPartsCtl)), pipeline.ForkBranch(pipeline.ServiceOperator(appStorageProvider)))), pipeline.WireSyncOperator("admin endpoint", adminEndpoint), pipeline.WireSyncOperator("bootstrap", bootstrapSyncOp), pipeline.WireSyncOperator("public endpoint", publicEndpoint),
+	)
 }
